python/requirements-api.txt
fastapi>=0.128.0
uvicorn>=0.42.0
numpy>=1.24.0
Pillow>=9.0.0

